+++ processor/src/routes/stripe-customer.route.ts
@@ -10,7 +10,7 @@ type CustomerRoutesOptions = {
 };
 
 export const customerRoutes = async (fastify: FastifyInstance, opts: FastifyPluginOptions & CustomerRoutesOptions) => {
-  fastify.get<{ Querystring: { stripeCustomerId?: string }; Reply: CustomerResponseSchemaDTO }>(
+  fastify.get<{ Querystring: { stripeCustomerId?: string }; Reply: CustomerResponseSchemaDTO | null }>(
     '/customer/session',
     {
       preHandler: [opts.sessionHeaderAuthHook.authenticate()],
@@ -31,7 +31,7 @@ export const customerRoutes = async (fastify: FastifyInstance, opts: FastifyPlug
       const { stripeCustomerId } = request.query;
       const resp = await opts.customerService.getCustomerSession(stripeCustomerId);
       if (!resp) {
-        return reply.status(204).send(resp);
+        return reply.status(204).send(null as any);
       }
       return reply.status(200).send(resp);
     },